requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
boto3==1.37.1
botocore==1.37.1
fastapi==0.115.8
httpx==0.28.1
mysql-connector-python==9.2.0
passlib==1.7.4
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.45.3
uvicorn==0.34.0
websockets==15.0
